+++ src/libb64/cdecode.c
@@ -1,9 +1,5 @@
 /****************************************************************************************************************************
-  cdecode.c - c source to a base64 decoding algorithm implementation
-  
-  This is part of the libb64 project, and has been placed in the public domain.
-  For details, see http://sourceforge.net/projects/libb64
-
+  cdecoder.c - c source to a base64 decoding algorithm implementation
   For STM32 with built-in Ethernet LAN8742A (Nucleo-144, DISCOVERY, etc) or W5x00/ENC28J60 shield/module
 
   EthernetWebServer_STM32 is a library for the STM32 running Ethernet WebServer
@@ -16,7 +12,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.3.1
+  Version: 1.3.2
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -34,25 +30,27 @@
   1.2.1   K Hoang      04/10/2021 Change option for PIO `lib_compat_mode` from default `soft` to `strict`. Update Packages Patches
   1.3.0   K Hoang      20/12/2021 Reduce usage of Arduino String with std::string. Use reference passing instead of value-passing
   1.3.1   K Hoang      25/12/2021 Fix bug
- *************************************************************************************************************************************/
+  1.3.2   K Hoang      28/12/2021 Fix wrong http status header bug and authenticate issue caused by libb64
+ *****************************************************************************************************************************/
 
 #include "cdecode.h"
 
 int base64_decode_value(char value_in)
 {
-  static const char decoding[] = {62, -1, -1, -1, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -2, -1, -1, -1, 0, 1, 2,
-                                  3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1,
-                                  -1, -1, -1, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46,
-                                  47, 48, 49, 50, 51
-                                 };
+  static const char decoding[] =
+  { 62, -1, -1, -1, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -2, -1, -1, -1, 0, 1, 2,
+    3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1,
+    -1, -1, -1, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46,
+    47, 48, 49, 50, 51
+  };
+
   static const char decoding_size = sizeof(decoding);
+  value_in -= 43;
 
-  int newValue = (int) value_in - 43;
-
-  if (newValue < 0 || newValue > decoding_size) 
+  if (value_in < 0 || value_in > decoding_size)
     return -1;
-    
-  return decoding[newValue];
+
+  return decoding[(int)value_in];
 }
 
 void base64_init_decodestate(base64_decodestate* state_in)
@@ -61,13 +59,11 @@ void base64_init_decodestate(base64_decodestate* state_in)
   state_in->plainchar = 0;
 }
 
-
 int base64_decode_block(const char* code_in, const int length_in, char* plaintext_out, base64_decodestate* state_in)
 {
   const char* codechar = code_in;
   char* plainchar = plaintext_out;
-  //char fragment;
-  int fragment;
+  char fragment;
 
   *plainchar = state_in->plainchar;
 
@@ -85,12 +81,10 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
             return plainchar - plaintext_out;
           }
 
-          fragment = base64_decode_value(*codechar++);
+          fragment = (char)base64_decode_value(*codechar++);
         } while (fragment < 0);
 
         *plainchar    = (fragment & 0x03f) << 2;
-
-        break;
 
       case step_b:
         do
@@ -102,13 +96,11 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
             return plainchar - plaintext_out;
           }
 
-          fragment = base64_decode_value(*codechar++);
+          fragment = (char)base64_decode_value(*codechar++);
         } while (fragment < 0);
 
         *plainchar++ |= (fragment & 0x030) >> 4;
         *plainchar    = (fragment & 0x00f) << 4;
-
-        break;
 
       case step_c:
         do
@@ -120,13 +112,11 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
             return plainchar - plaintext_out;
           }
 
-          fragment = base64_decode_value(*codechar++);
+          fragment = (char)base64_decode_value(*codechar++);
         } while (fragment < 0);
 
         *plainchar++ |= (fragment & 0x03c) >> 2;
         *plainchar    = (fragment & 0x003) << 6;
-
-        break;
 
       case step_d:
         do
@@ -138,22 +128,19 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
             return plainchar - plaintext_out;
           }
 
-          fragment = base64_decode_value(*codechar++);
+          fragment = (char)base64_decode_value(*codechar++);
         } while (fragment < 0);
 
         *plainchar++   |= (fragment & 0x03f);
-
-        break;
       }
-
   }
 
   /* control should not reach here */
   return plainchar - plaintext_out;
 }
 
-int base64_decode_chars(const char* code_in, const int length_in, char* plaintext_out)
-{
+int base64_decode_chars(const char* code_in, const int length_in, char* plaintext_out) {
+
   base64_decodestate _state;
   base64_init_decodestate(&_state);
   int len = base64_decode_block(code_in, length_in, plaintext_out, &_state);